fix: notifications are supposed to be on the landing page
